move to dedicated file, more operations, rename Comparable to SameType
