+++ src/crab/ebpf_domain.hpp
@@ -57,7 +57,7 @@ class ebpf_domain_t final {
     void operator()(const Assume&);
     void operator()(const Bin&);
     void operator()(const Call&);
-    void operator()(const Comparable&);
+    void operator()(const SameType&);
     void operator()(const Exit&);
     void operator()(const Jmp&);
     void operator()(const LoadMapFd&);
